+++ src/items/items.h
@@ -62,7 +62,7 @@ struct Abilities {
 		}
 
 		uint32_t getHealthGain() const {
-			return healthGain * static_cast<uint32_t>(g_configManager().getFloat(RATE_HEALTH_REGEN));
+			return healthGain * g_configManager().getFloat(RATE_HEALTH_REGEN);
 		}
 
 		void setHealthTicks(uint32_t value) {
@@ -70,7 +70,7 @@ struct Abilities {
 		}
 
 		uint32_t getHealthTicks() const {
-			return healthTicks / static_cast<uint32_t>(g_configManager().getFloat(RATE_HEALTH_REGEN_SPEED));
+			return healthTicks / g_configManager().getFloat(RATE_HEALTH_REGEN_SPEED);
 		}
 
 		void setManaGain(uint32_t value) {
@@ -78,7 +78,7 @@ struct Abilities {
 		}
 
 		uint32_t getManaGain() const {
-			return manaGain * static_cast<uint32_t>(g_configManager().getFloat(RATE_MANA_REGEN));
+			return manaGain * g_configManager().getFloat(RATE_MANA_REGEN);
 		}
 
 		void setManaTicks(uint32_t value) {
@@ -86,7 +86,7 @@ struct Abilities {
 		}
 
 		uint32_t getManaTicks() const {
-			return manaTicks / static_cast<uint32_t>(g_configManager().getFloat(RATE_MANA_REGEN_SPEED));
+			return manaTicks / g_configManager().getFloat(RATE_MANA_REGEN_SPEED);
 		}
 
 	private:
